requirements.txt

python-telegram-bot
python-dotenv

python-binance
pandas
numpy

pandas-ta

pytz